pattern

+++ circulradius.java
@@ -7,7 +7,7 @@ public class circulradius {
         Scanner sc=new Scanner(System.in);
         System.out.println("Enter the  radius : ");
         int r=sc.nextInt();
-        float area = (float)(3.14159 *r*r);
+        double area = (3.14159 *r*r);
         System.out.println("Area = "+area);
     }
 }
